add history button with last 10 successful inputs
Also tries to fix ctrl key getting stuck sometimes. And implemenets a
fix for outputting values being successful when window was change in
between inputting all the values. Now orange items that have been
changed will not turn back to blue unless a successful ctrl-1 is
outputted, in addition to the history not updating.